feat(dsl): 画面値の読み出し・Sel の対称化・秒の小数化とスレッド違反の非致命化
Swift DSL の評価で挙がった4点に対応する。

- `exist` の戻り値から `.text`/`.value`/`.id` を読めるようにする。新コマンドは足さず、
  `exist` が既に解決した要素を持ち帰るだけなので追加のデバイス往復もステップ記録も無い。
  値は成功時のみ載せる(掴めなかったのに値が読める状態を作らない)
- セレクタを取るコマンドの String / Sel を1対1にする。Shirates 由来の別名族
  (tapWithScrollDown 等8個)に Sel 版が無く、型付き経路を選ぶと機能が減る状態だった
- 秒引数を全て Double にする(`timeout: 1.2` が書ける)。表示は FTSeconds.format に集約し、
  ifCanSelect のポーリングも 0.5 秒固定をやめてサブ秒が丸められないようにする
- DSL スレッド外からの呼び出しで fatalError せず、失敗ステップを1回だけ記録して中断する。
  1プロセス=1シナリオなのでここで落とすとレポートごと消えていた。
  対になる排他(record/中断フラグ・FTRuntime の core 参照)を入れ、正しさは TSan で確認した

実行プロファイルの defaultTimeout が小数を取り得るようになるため、旧拡張が正当な値を
弾かないようプロトコル版を +1 する。

+++ Sources/FTCore/Flow.swift
@@ -41,9 +41,9 @@ public struct FlowStep: Codable, Sendable {
     public var direction: String?
     /// screenMatches 用の期待状態(自然言語。マルチモーダル画面検証に使う)
     public var expected: String?
-    /// 秒(整数)。検証(assert)では要素出現待ちの上限。アクションではロケータ解決の
+    /// 秒(小数可)。検証(assert)では要素出現待ちの上限。アクションではロケータ解決の
     /// 再試行待ち上限(nil = 既定の約0.7秒 / 0 = 再試行なし。optional ステップの空振り短縮用)
-    public var timeout: Int?
+    public var timeout: Double?
     /// scrollTo のスクロール回数上限(省略時 8)
     public var maxSwipes: Int?
     /// press の長押し秒数(nil = defaultPressDuration)。**ブリッジの /press は元から duration を
@@ -76,7 +76,7 @@ public struct FlowStep: Codable, Sendable {
 
     public init(action: String? = nil, assert: String? = nil, locator: FlowLocator? = nil,
                 fallbacks: [FlowLocator]? = nil, text: String? = nil, direction: String? = nil,
-                expected: String? = nil, timeout: Int? = nil, maxSwipes: Int? = nil,
+                expected: String? = nil, timeout: Double? = nil, maxSwipes: Int? = nil,
                 duration: Double? = nil,
                 expectedCount: Int? = nil,
                 optional: Bool? = nil, note: String? = nil, occlusionGuard: Bool? = nil) {
